+++ Source/Projects/PyPorts/Fibo/Fibo.hs
@@ -28,6 +28,6 @@ program Fibo
         
     {
         FiboDelegate fiboDelegate = Fibo;
-        Benchmark("Fibo", fiboDelegate, 24, 5); // Fibo(24) = 46368
+        Benchmark("Fibo", fiboDelegate, long(24), 5); // Fibo(24) = 46368
     }
 }
